Move `deployment` to sub-package and group tests into unit/integration
