--- dask_remote/__init__.py
@@ -1,1 +1,0 @@
-from .remote import DeploymentCluster  # noqa
