+++ dask_remote/deployment/__init__.py
@@ -1,0 +1,1 @@
+from .deployment import DeploymentCluster  # noqa
