+++ tests/test_integration/test_deployment/test_sync.py
@@ -2,7 +2,7 @@ import time
 
 import pytest
 
-from dask_remote import DeploymentCluster
+from dask_remote.deployment import DeploymentCluster
 
 
 @pytest.fixture
